+++ src/api/Models/IIdentifyable.cs
@@ -1,4 +1,6 @@
-﻿namespace BDMS.Models;
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace BDMS.Models;
 
 /// <summary>
 /// An object that can be identified by a numeric ID.
@@ -8,5 +10,6 @@ public interface IIdentifyable
     /// <summary>
     /// Gets or sets the entity's id.
     /// </summary>
+    [Required]
     public int Id { get; set; }
 }