+++ src/client/src/pages/detail/attachments/tabs/photo.ts
@@ -1,24 +1,10 @@
 import { useCallback } from "react";
 import { useQuery, useQueryClient, UseQueryResult } from "@tanstack/react-query";
-import { ApiError, NullableDateString, User } from "../../../../api/apiInterfaces.ts";
+import { ApiError } from "../../../../api/apiInterfaces.ts";
 import { download } from "../../../../api/download.ts";
 import { fetchApiV2Base, fetchApiV2Legacy, fetchApiV2WithApiError, upload } from "../../../../api/fetchApiV2.ts";
+import { Photo } from "../../../../api/generated";
 import { getImageFromBlob } from "../../../../utils.ts";
-
-export interface Photo {
-  id: number;
-  boreholeId: number;
-  name: string;
-  nameUuid: string;
-  fileType: string;
-  fromDepth: number;
-  toDepth: number;
-  public: boolean;
-  createdBy?: User;
-  created?: NullableDateString;
-  updatedBy?: User;
-  updated?: NullableDateString;
-}
 
 export const uploadPhoto = async (boreholeId: number, file: File): Promise<Photo> => {
   const formData = new FormData();