+++ src/client/src/pages/detail/detailHeader.tsx
@@ -4,13 +4,13 @@ import { Stack, Typography } from "@mui/material";
 import { ArrowDownToLine, Trash2, X } from "lucide-react";
 import { WorkflowStatus } from "@swissgeol/ui-core";
 import {
-  BoreholeV2,
   exportCSVBorehole,
   exportJsonBoreholes,
   exportJsonWithAttachmentsBorehole,
   useBoreholeEditable,
   useBoreholeMutations,
 } from "../../api/borehole.ts";
+import { Borehole } from "../../api/generated";
 import { useCurrentUser } from "../../api/user.ts";
 import { useAuth } from "../../auth/useBoreholesAuth.tsx";
 import {
@@ -31,7 +31,7 @@ import { SaveContext, SaveContextProps } from "./saveContext.tsx";
 import { StatusBadges } from "./statusBadges.tsx";
 
 interface DetailHeaderProps {
-  borehole: BoreholeV2;
+  borehole: Borehole;
 }
 
 const DetailHeader = ({ borehole }: DetailHeaderProps) => {