+++ src/client/src/pages/detail/form/hydrogeology/fieldMeasurement/fieldMeasurementDisplay.tsx
@@ -1,25 +1,27 @@
 import { FC } from "react";
 import { useTranslation } from "react-i18next";
 import { TableCell } from "@mui/material";
-import { Codelist, useCodelists } from "../../../../../components/codelist.ts";
+import { Codelist, FieldMeasurement, FieldMeasurementResult } from "../../../../../api/generated";
+import { useCodelistLocalizedLabel, useCodelists } from "../../../../../components/codelist.ts";
 import { DataDisplayCard } from "../../../../../components/dataCard/dataDisplayCard.tsx";
 import { FormResultTableDisplay } from "../../../../../components/form/formResultTableDisplay.js";
 import { parameterTableHeaderStyles } from "../../../../../components/form/formResultTableDisplayStyles.js";
 import { formatNumberForDisplay } from "../../../../../components/form/formUtils.ts";
 import ObservationDisplay from "../observationDisplay.tsx";
 import { getFieldMeasurementParameterUnits } from "../parameterUnits.js";
-import { deleteFieldMeasurement, FieldMeasurement, FieldMeasurementResult } from "./FieldMeasurement.ts";
+import { deleteFieldMeasurement } from "./FieldMeasurement.ts";
 
 export const FieldMeasurementDisplay: FC<{ item: FieldMeasurement }> = ({ item }) => {
   const codelists = useCodelists();
-  const { t, i18n } = useTranslation();
+  const { t } = useTranslation();
+  const getCodelistLabel = useCodelistLocalizedLabel();
 
   return (
     <DataDisplayCard<FieldMeasurement> item={item} deleteData={deleteFieldMeasurement} entityName={"fieldMeasurement"}>
       <ObservationDisplay observation={item} />
       <FormResultTableDisplay<FieldMeasurementResult>
         title={t("fieldMeasurementResult")}
-        results={item?.fieldMeasurementResults}
+        results={item?.fieldMeasurementResults ?? []}
         renderHeader={styles => (
           <>
             <TableCell sx={{ ...styles, paddingRight: 0 }}>{t("fieldMeasurementSampleType")}</TableCell>
@@ -30,7 +32,7 @@ export const FieldMeasurementDisplay: FC<{ item: FieldMeasurement }> = ({ item }
         renderBody={(result, index, styles) => (
           <>
             <TableCell sx={styles} data-cy={`fieldMeasurementResult.${index}.sampleType-formDisplay`}>
-              {codelists.data?.find((d: Codelist) => d.id === result.sampleTypeId)?.[i18n.language] ?? ""}
+              {getCodelistLabel(codelists.data?.find((d: Codelist) => d.id === result.sampleTypeId))}
             </TableCell>
             <TableCell
               component="th"
@@ -40,10 +42,10 @@ export const FieldMeasurementDisplay: FC<{ item: FieldMeasurement }> = ({ item }
                 ...parameterTableHeaderStyles,
               }}
               data-cy={`fieldMeasurementResult.${index}.parameter-formDisplay`}>
-              {codelists.data?.find((d: Codelist) => d.id === result.parameterId)?.[i18n.language] ?? ""}
+              {getCodelistLabel(codelists.data?.find((d: Codelist) => d.id === result.parameterId))}
             </TableCell>
             <TableCell sx={styles} data-cy={`fieldMeasurementResult.${index}.value-formDisplay`}>
-              {result?.value && result?.parameterId && (
+              {result?.value != null && result?.parameterId != null && (
                 <>
                   <span>{formatNumberForDisplay(result?.value) + " "}</span>
                   {getFieldMeasurementParameterUnits(result.parameterId, codelists.data)}