+++ CHANGELOG.md
@@ -14,6 +14,7 @@
 - Reference Codelist instead of enum for OriginalReferenceSystem in borehole.
 - Lithology rock type can now be left unspecified (in addition to unconsolidated and consolidated). Stratigraphy extraction creates lithologies as unspecified by default.
 - Redesigned lithology table to edit depths in the table overview and stratigraphy extraction.
+- Typescript types for backend models are now generated from swagger doc.
 
 ### Fixed
 - Fixed bug where casing name was not correctly displayed in the select on the backfill card.